Fixed close of server for client tests * Moved `runServer` function to lib * Added server init to `runServer` function * Added server close to `runServer` function
