+++ test/techs/bh-client-module.test.js
@@ -1,10 +1,6 @@
 var fs = require('fs'),
     path = require('path'),
     mock = require('mock-fs'),
-    promisify = require('vow-node').promisify,
-    http = require('http'),
-    serveStatic = require('serve-static'),
-    finalhandler = require('finalhandler'),
     TestNode = require('enb/lib/test/mocks/test-node'),
     FileList = require('enb/lib/file-list'),
     bhClientModule = require('../../techs/bh-client-module'),
@@ -14,20 +10,9 @@ var fs = require('fs'),
     chaiFilename = require.resolve('chai/chai.js'),
     ymFilename = require.resolve('ym/modules.js'),
     writeFile = require('../lib/write-file'),
-    runPhantom = require('../lib/run-phantom'),
-    serve, server, listen;
+    runServer = require('../lib/run-server');
 
 describe('bh-client-module', function () {
-    beforeEach(function () {
-        serve = serveStatic(process.cwd(), { index: false });
-        server = http.createServer(function (req, res) {
-            var done = finalhandler(req, res);
-
-            serve(req, res, done);
-        });
-        listen = promisify(server.listen.bind(server));
-    });
-
     afterEach(function () {
         mock.restore();
     });
@@ -117,31 +102,6 @@ describe('bh-client-module', function () {
        return runTest(test, options, template, lib);
     });
 
-    it('sourcemap', function () {
-        var options = {
-                sourcemap: true,
-                bhFile: 'bh.js'
-            },
-            scheme = {
-                blocks: {},
-                bundle: {},
-                'bh.js': 'module.exports = BH;'
-            },
-            bundle, fileList;
-
-        mock(scheme);
-
-        bundle = new TestNode('bundle');
-        fileList = new FileList();
-        fileList.loadFromDirSync('blocks');
-        bundle.provideTechData('?.files', fileList);
-
-        return bundle.runTechAndGetContent(bhClientModule, options)
-            .spread(function (bh) {
-                bh.toString().must.include('sourceMappingURL');
-            });
-    });
-
     describe('caches', function () {
         var mockBhCore, scheme, bundle, fileList;
 
@@ -160,10 +120,6 @@ describe('bh-client-module', function () {
                 'chai.js': fs.readFileSync(chaiFilename, 'utf-8'),
                 'ym.js': fs.readFileSync(ymFilename, 'utf-8')
             };
-        });
-
-        afterEach(function () {
-            server.close();
         });
 
         it('must use cached bhFile', function () {
@@ -196,16 +152,10 @@ describe('bh-client-module', function () {
                     return bundle.runTechAndGetContent(bhClientModule, { bhFile: 'mock.bh.js' });
                 })
                 .spread(function (bh) {
-                    fs.writeFileSync('bundle/bundle.bh.js', bh);
                     // TODO: удалить, когда пофиксится https://github.com/enb-make/enb/issues/224
+                    fs.writeFileSync('bundle/bundle.bh.js', bh);
 
-                    return listen(3000);
-                })
-                .then(function () {
-                    return runPhantom('http://localhost:3000/index.html');
-                })
-                .fail(function (err) {
-                    throw err;
+                    return runServer(3000);
                 });
         });
 
@@ -239,16 +189,10 @@ describe('bh-client-module', function () {
                     return bundle.runTechAndGetContent(bhClientModule, { bhFile: 'mock.bh.js' });
                 })
                 .spread(function (bh) {
-                    fs.writeFileSync('bundle/bundle.bh.js', bh);
                     // TODO: удалить, когда пофиксится https://github.com/enb-make/enb/issues/224
+                    fs.writeFileSync('bundle/bundle.bh.js', bh);
 
-                    return listen(3000);
-                })
-                .then(function () {
-                    return runPhantom('http://localhost:3000/index.html');
-                })
-                .fail(function (err) {
-                    throw err;
+                    return runServer(3000);
                 });
         });
 
@@ -279,16 +223,10 @@ describe('bh-client-module', function () {
                     return bundle.runTechAndGetContent(bhClientModule);
                 })
                 .spread(function (bh) {
-                    fs.writeFileSync('bundle/bundle.bh.js', bh);
                     // TODO: удалить, когда пофиксится https://github.com/enb-make/enb/issues/224
+                    fs.writeFileSync('bundle/bundle.bh.js', bh);
 
-                    return listen(3000);
-                })
-                .then(function () {
-                    return runPhantom('http://localhost:3000/index.html');
-                })
-                .fail(function (err) {
-                    throw err;
+                    return runServer(3000);
                 });
         });
     });
@@ -335,19 +273,10 @@ function runTest(testContent, options, template, lib) {
 
     return bundle.runTechAndGetContent(bhClientModule, options)
         .spread(function (bh) {
-            fs.writeFileSync('bundle/bundle.bh.js', bh);
             // TODO: удалить, когда пофиксится https://github.com/enb-make/enb/issues/224
+            fs.writeFileSync('bundle/bundle.bh.js', bh);
 
-            return listen(3000);
-        })
-        .then(function () {
-            return runPhantom('http://localhost:3000/index.html');
-        })
-        .then(function () {
-            server.close();
-        })
-        .fail(function (err) {
-            throw err;
+            return runServer(3000);
         });
 }
 
@@ -364,4 +293,3 @@ function generateTest(json, result) {
             '});'
     ].join('\n');
 }
-
